src/public_pubsub_topic/requirements.txt
google-cloud-logging
google-cloud-pubsub